requirements.txt
Pillow>=9.0.0
requests>=2.25.0
huggingface_hub
python-dotenv>=1.0.0
ollama>=0.1.0